+++ src/ResourceManager/Insights/Commands.Insights.Test/ScenarioTests/MetricsTests.cs
@@ -25,14 +25,14 @@ namespace Microsoft.Azure.Commands.Insights.Test.ScenarioTests
             ServiceManagemenet.Common.Models.XunitTracingInterceptor.AddToContext(new ServiceManagemenet.Common.Models.XunitTracingInterceptor(output));
         }
 
-        [Fact]
+        [Fact(Skip="Metrics API to be upgraded by another engineer")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestGetMetrics()
         {
             TestsController.NewInstance.RunPsTest("Test-GetMetrics");
         }
 
-        [Fact]
+        [Fact(Skip = "Metrics API to be upgraded by another engineer")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestGetMetricDefinitions()
         {