+++ src/ResourceManager/Insights/Commands.Insights/Alerts/NewAzureRmAlertRuleWebhookCommand.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Commands.Insights.Alerts
     /// <summary>
     /// Create an AlertRuleWebhook action
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmAlertRuleWebhook"), OutputType(typeof(RuleWebhookAction))]
+    [Cmdlet(VerbsCommon.New, "AzureRmAlertRuleWebhook"), OutputType(typeof(Management.Monitor.Management.Models.RuleWebhookAction))]
     public class NewAzureRmAlertRuleWebhookCommand : MonitorCmdletBase
     {
         /// <summary>
@@ -55,7 +55,7 @@ namespace Microsoft.Azure.Commands.Insights.Alerts
                 ? new Dictionary<string, string>()
                 : this.Property.Keys.Cast<object>().ToDictionary(key => (string)key, key => (string)this.Property[key]);
 
-            var action = new RuleWebhookAction
+            var action = new Management.Monitor.Management.Models.RuleWebhookAction
             {
                 ServiceUri = this.ServiceUri,
                 Properties = dictionary