+++ src/ResourceManager/Insights/Commands.Insights/Metrics/GetAzureRmMetricDefinitionCommand.cs
@@ -61,10 +61,12 @@ namespace Microsoft.Azure.Commands.Insights.Metrics
                 cmdletName: cmdletName,
                 topic: "Parameter deprecation",
                 message: "The DetailedOutput parameter will be deprecated in a future breaking change release.");
+
             this.WriteIdentifiedWarning(
                 cmdletName: cmdletName,
                 topic: "Parameter name change", 
                 message: "The parameter plural names for the parameters will be deprecated in a future breaking change release in favor of the singular versions of the same names.");
+
             bool fullDetails = this.DetailedOutput.IsPresent;
 
             // Get metricDefintions and filter the response to return metricDefinitions for only the specified metric names